handle routing/scrolling mayhem
